requirements.txt
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
#dlib==19.22.1
#dlib /D:/FRA/backend/dlib-19.22.99-cp310-cp310-win_amd64.whl#sha256=f2181b6724669fb0147c9ebc326d72b94de41bced1812eff5df4b133b5d0b575
dnspython==2.7.0
#face-recognition==1.3.0
#face_recognition_models==0.3.0
gunicorn==23.0.0
numpy==1.26.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.1.0
pymongo==4.12.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
cmake==3.31.4